A data frame of the documentation of pharmaverse data

A data frame of the documentation of pharmaverse data

Usage

pharmadverse

Format

An object of class tbl_df (inherits from tbl, data.frame) with 74 rows and 7 columns.

Examples

pharmadverse
#> # A tibble: 74 × 7
#>    topic                               alias title concept type  keyword package
#>    <chr>                               <chr> <chr> <chr>   <chr> <chr>   <chr>  
#>  1 <a href=https://pharmaverse.github… adae  adae  Datase… help  datase… pharma…
#>  2 <a href=https://pharmaverse.github… adbc… adbc… Datase… help  datase… pharma…
#>  3 <a href=https://pharmaverse.github… adce… adce… Datase… help  datase… pharma…
#>  4 <a href=https://pharmaverse.github… adcm  adcm  Datase… help  datase… pharma…
#>  5 <a href=https://pharmaverse.github… adeg  adeg  Datase… help  datase… pharma…
#>  6 <a href=https://pharmaverse.github… adex  adex  Datase… help  datase… pharma…
#>  7 <a href=https://pharmaverse.github… adfa… adfa… Datase… help  datase… pharma…
#>  8 <a href=https://pharmaverse.github… adis… adis… Datase… help  datase… pharma…
#>  9 <a href=https://pharmaverse.github… adlb  adlb  Datase… help  datase… pharma…
#> 10 <a href=https://pharmaverse.github… adlb… adlb… Datase… help  datase… pharma…
#> # ℹ 64 more rows